+++ examples/priority_search_on_convex_fn.py
@@ -212,7 +212,7 @@ import numpy as np
 from opto import trace
 from opto.utils.llm import LLM, LiteLLM
 from opto.optimizers import OptoPrimeV2 as OptoPrime
-from opto.features.priority_search import PrioritySearch as SearchAlgorithm
+from opto.trainer.algorithms import PrioritySearch as SearchAlgorithm
 from opto.trainer.guide import Guide
 from opto.trainer.loggers import TensorboardLogger
 from opto.trainer.guide import LLMJudge
