+++ opto/trainer/algorithms/__init__.py
@@ -1,4 +1,4 @@
 from opto.trainer.algorithms.algorithm import Trainer
-from opto.trainer.algorithms.basic_algorithms import Minibatch, MinibatchAlgorithm, BasicSearchAlgorithm
-from opto.trainer.algorithms.beamsearch_algorithm import BeamsearchAlgorithm, BeamsearchHistoryAlgorithm
-from opto.trainer.algorithms.UCBsearch import UCBSearchAlgorithm
+from opto.trainer.algorithms.priority_search import PrioritySearch
+from opto.trainer.algorithms.polca import POLCA
+from opto.trainer.algorithms.streaming_priority_search import StreamingPrioritySearch
