+++ opto/trainer/examples/__init__.py
@@ -1,0 +1,3 @@
+from opto.trainer.examples.basic_algorithms import Minibatch, MinibatchAlgorithm, BasicSearchAlgorithm
+from opto.trainer.examples.beamsearch_algorithm import BeamsearchAlgorithm, BeamsearchHistoryAlgorithm
+from opto.trainer.examples.UCBsearch import UCBSearchAlgorithm
