+++ opto/trainer/examples/aggregator.py
@@ -8,8 +8,7 @@ from typing import Dict, List, Any, Union
 from opto import trace
 from opto.trace.nodes import ParameterNode
 from opto.optimizers.utils import print_color
-from opto.trainer.algorithms import Minibatch
-from opto.trainer.algorithms.basic_algorithms import standard_optimization_step
+from opto.trainer.examples.basic_algorithms import Minibatch, standard_optimization_step
 from opto.utils.llm import LLM, AbstractModel
