+++ opto/trainer/examples/basic_algorithms.py
